requirements.txt
flask==3.1.0
flask-cors==5.0.1
anthropic==0.50.0
Pillow==11.2.1
gunicorn==23.0.0
